fix aggro cuz bivins is a gigantic doofus idiot dumbo turdburgler weiner sniffer nincompoop gosh darnit

+++ Chaos/Scripting/Components/AbilityComponents/AddAggroComponent.cs
@@ -27,7 +27,7 @@ public class AddAggroComponent : IComponent
 
             monster?.AggroList.AddOrUpdate(
                 context.Source.Id,
-                _ => aggroAmount * (int)aggroMultiplier,
+                _ => aggroAmount * aggroMultiplier,
                 (_, currentAggro) => currentAggro + aggroAmount * aggroMultiplier);
         }
     }
